[practice-problems] chore: Add find-longest-bitonic-subarray-array
Find Longest Bitonic Subarray in an array.
The longest bitonic subarray problem is to find a subarray of a given sequence in which the subarray’s elements are first sorted in in increasing order, then in decreasing order, and the subarray is as long as possible. Strictly ascending or descending subarrays are also accepted.